Dynamic control board selection.
With this commit, the user can dynamically select the control board they want to use.
This makes the act of switching between the real board and the simulator much faster.

Unfortunatly, there does seem to be some instability with this commit. It's fine as
long as you don't switch between control board interfaces a bunch.

Another issue I'm seeing is that Test Mode's controls sometimes appear in the top-left
corner of the GUI. Scrolling the window seems to get the controls in the place they
need to be.